refactor: batch 모듈 추가
