+++ offroad-api/src/main/java/site/offload/api/character/service/GainedCharacterService.java
@@ -7,8 +7,6 @@ import site.offload.db.character.entity.GainedCharacterEntity;
 import site.offload.db.character.repository.GainedCharacterRepository;
 import site.offload.db.member.entity.MemberEntity;
 
-import java.util.List;
-
 @Component
 @RequiredArgsConstructor
 public class GainedCharacterService {
