+++ offroad-api/src/main/java/site/offload/api/quest/service/QuestRewardService.java
@@ -3,7 +3,6 @@ package site.offload.api.quest.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import site.offload.api.exception.NotFoundException;
-import site.offload.db.quest.entity.QuestEntity;
 import site.offload.db.quest.entity.QuestRewardEntity;
 import site.offload.db.quest.repository.QuestRewardRepository;
 import site.offload.enums.response.ErrorMessage;
