+++ offroad-batch/build.gradle
@@ -1,0 +1,14 @@
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+
+    implementation project(path: ':offroad-enum')
+    implementation project(path: ':offroad-db')
+}
+
+bootJar {
+    enabled = false
+}
+
+jar {
+    enabled = true
+}
