+++ offroad-db/src/main/java/site/offload/db/character/repository/GainedCharacterRepository.java
@@ -6,9 +6,6 @@ import site.offload.db.character.entity.CharacterEntity;
 import site.offload.db.character.entity.GainedCharacterEntity;
 import site.offload.db.member.entity.MemberEntity;
 
-import java.lang.reflect.Member;
-import java.util.List;
-
 public interface GainedCharacterRepository extends JpaRepository<GainedCharacterEntity, Long> {
 
     boolean existsByCharacterEntityAndMemberEntity(CharacterEntity characterEntity, MemberEntity memberEntity);
