+++ offroad-db/src/main/java/site/offload/db/emblem/repository/GainedEmblemRepository.java
@@ -2,7 +2,6 @@ package site.offload.db.emblem.repository;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import site.offload.db.emblem.entity.GainedEmblemEntity;
 import site.offload.db.member.entity.MemberEntity;
