+++ settings.gradle
@@ -6,4 +6,5 @@ include 'offroad-external'
 include 'offroad-enum'
 include 'offroad-cache'
 include 'offroad-support'
+include 'offroad-batch'
